Add some extra modules to build ALE out of tree
The ALE core is pretty self-contained, but it depends on
MOM_error_handler (which we replace with a blank stub) and some string
functions (which we can copy directly). The MOM_memory.h header is
also blank, and required so the Fortran preprocessor will succeed.

We also ignore binary result files (.mod, .o, .so)